# docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_go-park-mail-ru_2026_1_TheBugs_internal_delivery_restapi_auth.LoginResponse:
    properties:
      access_token:
        type: string
      expire_at:
        type: integer
    type: object
  request.UpdatePwdDTO:
    properties:
      password:
        type: string
    type: object
  request.UserEmail:
    properties:
      email:
        type: string
    type: object
  request.VerifyCodeDTO:
    properties:
      code:
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      details:
        type: string
      error:
        type: string
    type: object
...
  response.ValidationErrorResponse:
    properties:
      error:
        type: string
      field:
        type: string
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Created by TheBugs in 2026
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: DomDeli API
  version: "1.0"
paths:
  /auth/admin/login:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Authenticate admin user and return access token + set refresh token
        cookie
      parameters:
      - description: User email
        in: formData
        name: email
        required: true
        type: string
      - description: User password
        in: formData
        name: password
        required: true
        type: string
      responses:
        "200":
          description: Successful login, returns access token
          headers:
            Set-Cookie:
              description: refresh_token=<NEW_REFRESH_TOKEN>; HttpOnly; Path=/api/auth/refresh;
                Max-Age=...
              type: string
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ValidationErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - CSRFToken: []
      summary: Login admin user
      tags:
      - Auth
  /auth/email:
    post:
      consumes:
      - application/json
      description: Generates recovery session, sends verification code to user's email
        and sets session_id cookie
      parameters:
      - description: User email
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.UserEmail'
      produces:
      - application/json
      responses:
        "200":
          description: Status OK
          headers:
            Set-Cookie:
              description: session_id=<SESSION_ID>; HttpOnly; Path=/api/auth; Max-Age=600
              type: string
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ValidationErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - CSRFToken: []
      summary: Send verify code to email
      tags:
      - Auth
  /auth/email/verify:
    post:
      consumes:
      - application/json
      description: Verifies code from email using session_id cookie and marks session
        as verified
      parameters:
      - description: Verification code
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.VerifyCodeDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Status verified
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ValidationErrorResponse'
        "401":
          description: Invalid code or session
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - CSRFToken: []
      summary: Verify user email from code
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Authenticate user and return access token + set refresh token cookie
      parameters:
      - description: User email
        in: formData
        name: email
        required: true
        type: string
      - description: User password
        in: formData
        name: password
        required: true
        type: string
      responses:
        "200":
          description: Successful login, returns access token
          headers:
            Set-Cookie:
              description: refresh_token=<NEW_REFRESH_TOKEN>; HttpOnly; Path=/api/auth/refresh;
                Max-Age=...
              type: string
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_TheBugs_internal_delivery_restapi_auth.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ValidationErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - CSRFToken: []
      summary: Login user
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Blacklist access/refresh tokens and delete refresh cookie
      produces:
      - application/json
      responses:
        "204":
          description: Successfully logged out (no content)
        "400":
          description: Missing tokens
          schema:
            $ref: '#/definitions/response.ValidationErrorResponse'
        "401":
          description: Invalid tokens
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Blacklist error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      - CSRFToken: []
      summary: User logout
      tags:
      - Auth
  /auth/recover:
    post:
      consumes:
      - application/json
      description: Generates recovery session, sends verification code to user's email
        and sets session_id cookie
      parameters:
      - description: User email
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.UserEmail'
      produces:
      - application/json
      responses:
        "200":
          description: Status OK
          headers:
            Set-Cookie:
              description: session_id=<SESSION_ID>; HttpOnly; Path=/api/auth; Max-Age=600
              type: string
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ValidationErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - CSRFToken: []
      summary: Send recovery code to email
      tags:
      - Auth
  /auth/recover/reset:
    post:
      consumes:
      - application/json
      description: Updates password using verified recovery session (session_id cookie
        required)
      parameters:
      - description: New password
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.UpdatePwdDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Password updated
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ValidationErrorResponse'
        "401":
          description: Session not verified or invalid
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - CSRFToken: []
      summary: Verify user email
      tags:
      - Auth
  /auth/recover/verify:
    post:
      consumes:
      - application/json
      description: Verifies code from email using session_id cookie and marks session
        as verified
      parameters:
      - description: Verification code
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.VerifyCodeDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Status verified
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ValidationErrorResponse'
        "401":
          description: Invalid code or session
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - CSRFToken: []
      summary: Verify recovery code
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Obtain new access token using refresh token stored in cookie (refresh_token
        cookie required)
      produces:
      - application/json
      responses:
        "200":
          description: New access token, also updates refresh token cookie
          headers:
            Set-Cookie:
              description: new refresh_token=...; HttpOnly; Path=/api/auth/refresh;
                Max-Age=...
              type: string
          schema:
            $ref: '#/definitions/github_com_go-park-mail-ru_2026_1_TheBugs_internal_delivery_restapi_auth.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ValidationErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - CSRFToken: []
      summary: Refresh access token
      tags:
      - Auth
...
  /csrf-token:
    get:
      consumes:
      - application/json
      description: Create and get csrf token
      produces:
      - application/json
      responses:
        "200":
          description: CSRF token
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get CSRF token
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
  CSRFToken:
    description: CSRF токен в заголовке X-CSRF-Token для защищённых
    in: header
    name: X-CSRF-Token
    type: apiKey
swagger: "2.0"